requirements.txt
Flask==3.0.3
pytest==8.2.0
flask_testing==0.8.1
flask-sqlalchemy==3.1.1
python-dotenv==1.0.1
psycopg2==2.9.9